LayoutTests/imported/w3c/web-platform-tests/intersection-observer/svg-stroke-change-expected.txt


PASS IntersectionObserver observing an SVG element while its stroke is changed
FAIL First rAF. assert_approx_equals: entries[0].boundingClientRect.right expected 108 +/- 0 but got 8
FAIL change stroke assert_equals: entries.length expected 2 but got 1
